Add KIND parameter to DECLARE_SHADER
The pattern followed for all other ShaderTrap commands is that every
parameter is named. For DECLARE_SHADER there is only one parameter -
the kind of the shader - and it had seemed a bit redundant to force it
to be named, but actually for consistency this seems like a good
idea. This change requires a shader to be declared using:

DECLARE_SHADER result KIND VERTEX|FRAGMENT|COMPUTE
// Text
END