Encapsulate GL functions via struct of function objects (#39)
To avoid libshadertrap committing to a specific source of GL
functions, this change introduces a struct of function objects with a
member for every OpenGL ES function. An Executor object is constructed
with an instance of this struct, allowing GL functions suitable for
the platform in question to be passed.