More general support for setting uniforms (#74)
Before this change, the SET_UNIFORM command could take an identifier
for the named uniform to be set. Now, SET_UNIFORM takes a string to
identify the uniform, which allows setting the values of struct
uniforms, and arrays of structs. The change includes an illustrative
example, and updates the language documentation.